Android - Prediction

+++ src/Android/app/src/main/java/com/example/dermaai_android_140/myClasses/Diagnosis.kt
@@ -1,3 +1,3 @@
 package com.example.dermaai_android_140.myClasses
 
-data class Diagnosis (val prediction: Map<String, Int>, val imagePath : String)
+data class Diagnosis (val prediction: Map<String, Double>, val imagePath : String)
